refactor: more tidy
